requirements.txt
asgiref==3.7.2
channels==4.0.0
channels-redis==4.2.0
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
msgpack==1.0.8
OSlash==0.6.3
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
redis==5.0.3
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
